+++ include/obsvfun.h
@@ -5,19 +5,19 @@
 
 template<class state_type, class obsv_type>
 class obsvfun :
-    public std::binary_function<state_type, obsv_type, double>
+    public std::binary_function<state_type, obsv_type, long double>
 {
     public:
         obsvfun();
-        obsvfun(double (*f)(state_type,obsv_type));
+        obsvfun(long double (*f)(state_type,obsv_type));
         virtual ~obsvfun();
         obsvfun(const obsvfun& other);
         obsvfun& operator=(const obsvfun& other);
 
-        double virtual operator() (const state_type a, const obsv_type b) const {  return f(a,b) ;}
+        long double virtual operator() (const state_type a, const obsv_type b) const {  return f(a,b) ;}
     protected:
     private:
-        double (*f)(state_type, obsv_type );
+        long double (*f)(state_type, obsv_type );
 
 
 };
@@ -32,7 +32,7 @@ obsvfun<state_type,obsv_type>::obsvfun()
 }
 
 template<class state_type, class obsv_type>
-obsvfun<state_type,obsv_type>::obsvfun(double (*f)(state_type,obsv_type))
+obsvfun<state_type,obsv_type>::obsvfun(long double (*f)(state_type,obsv_type))
 :f(f)
 {
